Changes form the Threema Message SDK 1.1.4

+++ src/main/java/com/neilalexander/jnacl/crypto/verify_16.java
@@ -1,17 +1,17 @@
 //
 //  Copyright (c) 2011, Neil Alexander T.
 //  All rights reserved.
-// 
+//
 //  Redistribution and use in source and binary forms, with
 //  or without modification, are permitted provided that the following
 //  conditions are met:
-// 
+//
 //  - Redistributions of source code must retain the above copyright notice,
 //    this list of conditions and the following disclaimer.
 //  - Redistributions in binary form must reproduce the above copyright notice,
 //    this list of conditions and the following disclaimer in the documentation
 //    and/or other materials provided with the distribution.
-// 
+//
 //  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 //  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 //  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -27,17 +27,14 @@
 
 package com.neilalexander.jnacl.crypto;
 
-public class verify_16
-{
-	final int crypto_verify_16_ref_BYTES = 16;
+public class verify_16 {
+  final int crypto_verify_16_ref_BYTES = 16;
 
-	public static int crypto_verify(byte[] x, int xoffset, byte[] y)
-	{
-		int differentbits = 0;
-		
-		for (int i = 0; i < 15; i++)
-			differentbits |= ((int)(x[xoffset + i] ^ y[i])) & 0xff;
-		
-		return (1 & (((int)differentbits - 1) >>> 8)) - 1;
-	}
+  public static int crypto_verify(byte[] x, int xoffset, byte[] y) {
+    int differentbits = 0;
+
+    for (int i = 0; i < 15; i++) differentbits |= (x[xoffset + i] ^ y[i]) & 0xff;
+
+    return (1 & ((differentbits - 1) >>> 8)) - 1;
+  }
 }
